Centralize derive-safe builtin names in internal/derivepure
Wire runtime builtins to derivepure for isBuiltinAllowedInDerive and remove
the duplicate pure_builtins map.
